sub-image/pyproject.toml
[tool.poetry]
name = "sub-image"
version = "0.1.0"
description = ""
authors = ["sevenc-nanashi <sevenc7c@sevenc7c.com>"]
readme = "README.md"
packages = [{ include = "sub_image" }]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.88.0"
pjsekai-background-gen-pillow = "^0.3.0"
requests = "^2.28.1"
uvicorn = "^0.20.0"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.0"
poethepoet = "^0.17.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
start = "gunicorn -w 4 -k uvicorn.workers.UvicornWorker -b 0.0.0.0:3203 main:app"
dev = "uvicorn main:app --reload --port 3203"
